refactor: move model file IO into dedicated module (#1442)
