+++ src/model_io/ckpt_io.cpp
@@ -1,0 +1,403 @@
+#include "ckpt_io.h"
+
+#include <cstdint>
+#include <cstdio>
+#include <cstdlib>
+#include <cstring>
+#include <string>
+#include <vector>
+
+#include "zip.h"
+
+static constexpr int MAX_STRING_BUFFER = 512;
+
+static void set_error(std::string* error, const std::string& message) {
+    if (error != nullptr) {
+        *error = message;
+    }
+}
+
+static int32_t read_int(const uint8_t* buffer) {
+    // little endian
+    uint32_t value = 0;
+    value |= static_cast<uint32_t>(buffer[3]) << 24;
+    value |= static_cast<uint32_t>(buffer[2]) << 16;
+    value |= static_cast<uint32_t>(buffer[1]) << 8;
+    value |= static_cast<uint32_t>(buffer[0]);
+    return static_cast<int32_t>(value);
+}
+
+static uint16_t read_short(const uint8_t* buffer) {
+    // little endian
+    uint16_t value = 0;
+    value |= static_cast<uint16_t>(buffer[1]) << 8;
+    value |= static_cast<uint16_t>(buffer[0]);
+    return value;
+}
+
+bool is_ckpt_file(const std::string& file_path) {
+    zip_t* zip = zip_open(file_path.c_str(), 0, 'r');
+    if (zip == nullptr) {
+        return false;
+    }
+    zip_close(zip);
+    return true;
+}
+
+/*================================================= CkptModelLoader ==================================================*/
+
+// $ python -m pickletools sd-v1-4/archive/data.pkl | head -n 100
+//     0: \x80 PROTO      2
+//     2: }    EMPTY_DICT
+//     3: q    BINPUT     0
+//     5: (    MARK
+//     6: X        BINUNICODE 'epoch'
+//    16: q        BINPUT     1
+//    18: K        BININT1    6
+//    20: X        BINUNICODE 'global_step'
+//    36: q        BINPUT     2
+//    38: J        BININT     470000
+//    43: X        BINUNICODE 'pytorch-lightning_version'
+//    73: q        BINPUT     3
+//    75: X        BINUNICODE '1.4.2'
+//    85: q        BINPUT     4
+//    87: X        BINUNICODE 'state_dict'
+//   102: q        BINPUT     5
+//   104: }        EMPTY_DICT
+//   105: q        BINPUT     6
+//   107: (        MARK
+//   108: X            BINUNICODE 'betas'
+//   118: q            BINPUT     7
+//   120: c            GLOBAL     'torch._utils _rebuild_tensor_v2'
+//   153: q            BINPUT     8
+//   155: (            MARK
+//   156: (                MARK
+//   157: X                    BINUNICODE 'storage'
+//   169: q                    BINPUT     9
+//   171: c                    GLOBAL     'torch FloatStorage'
+//   191: q                    BINPUT     10
+//   193: X                    BINUNICODE '0'
+//   199: q                    BINPUT     11
+//   201: X                    BINUNICODE 'cpu'
+//   209: q                    BINPUT     12
+//   211: M                    BININT2    1000
+//   214: t                    TUPLE      (MARK at 156)
+//   215: q                BINPUT     13
+//   217: Q                BINPERSID
+//   218: K                BININT1    0
+//   220: M                BININT2    1000
+//  ...............................
+//  3201: q            BINPUT     250
+//  3203: R            REDUCE
+//  3204: q            BINPUT     251
+//  3206: X            BINUNICODE 'model.diffusion_model.input_blocks.1.1.proj_in.weight'
+//  3264: q            BINPUT     252
+//  3266: h            BINGET     8
+//  3268: (            MARK
+//  3269: (                MARK
+//  3270: h                    BINGET     9
+//  3272: h                    BINGET     10
+//  3274: X                    BINUNICODE '30'
+//  3281: q                    BINPUT     253
+//  3283: h                    BINGET     12
+//  3285: J                    BININT     102400
+//  3290: t                    TUPLE      (MARK at 3269)
+//  3291: q                BINPUT     254
+//  3293: Q                BINPERSID
+//  3294: K                BININT1    0
+//  3296: (                MARK
+//  3297: M                    BININT2    320
+//  3300: M                    BININT2    320
+//  3303: K                    BININT1    1
+//  3305: K                    BININT1    1
+//  3307: t                    TUPLE      (MARK at 3296)
+//  3308: q                BINPUT     255
+//  3310: (                MARK
+//  3311: M                    BININT2    320
+//  3314: K                    BININT1    1
+//  3316: K                    BININT1    1
+//  3318: K                    BININT1    1
+//  3320: t                    TUPLE      (MARK at 3310)
+//  3321: r                LONG_BINPUT 256
+//  3326: \x89             NEWFALSE
+//  3327: h                BINGET     16
+//  3329: )                EMPTY_TUPLE
+//  3330: R                REDUCE
+//  3331: r                LONG_BINPUT 257
+//  3336: t                TUPLE      (MARK at 3268)
+//  3337: r            LONG_BINPUT 258
+//  3342: R            REDUCE
+//  3343: r            LONG_BINPUT 259
+//  3348: X            BINUNICODE 'model.diffusion_model.input_blocks.1.1.proj_in.bias'
+//  3404: r            LONG_BINPUT 260
+//  3409: h            BINGET     8
+//  3411: (            MARK
+//  3412: (                MARK
+//  3413: h                    BINGET     9
+//  3415: h                    BINGET     10
+//  3417: X                    BINUNICODE '31'
+
+struct PickleTensorReader {
+    enum ReadPhase {
+        READ_NAME,
+        READ_DATA,
+        CHECK_SIZE,
+        READ_DIMENS
+    };
+    ReadPhase phase   = READ_NAME;
+    size_t entry_size = 0;
+    int32_t nelements = 0;
+
+    TensorStorage tensor_storage;
+
+    static ggml_type global_type;  // all pickle_tensors data type
+    static bool read_global_type;
+
+    bool read_int_value(uint32_t value) {
+        if (phase == CHECK_SIZE) {
+            if (entry_size == value * ggml_type_size(tensor_storage.type)) {
+                nelements = value;
+                phase     = READ_DIMENS;
+                return true;
+            } else {
+                phase = READ_NAME;
+            }
+        } else if (phase == READ_DIMENS) {
+            if (tensor_storage.n_dims + 1 > SD_MAX_DIMS) {  // too many dimens
+                phase                 = READ_NAME;
+                tensor_storage.n_dims = 0;
+            }
+            if (nelements % value == 0) {
+                tensor_storage.ne[tensor_storage.n_dims] = value;
+                tensor_storage.n_dims++;
+            }
+        }
+        return false;
+    }
+
+    void read_global(const std::string& str) {
+        if (str == "FloatStorage") {
+            if (read_global_type) {
+                global_type      = GGML_TYPE_F32;
+                read_global_type = false;
+            }
+            tensor_storage.type = GGML_TYPE_F32;
+        } else if (str == "HalfStorage") {
+            if (read_global_type) {
+                global_type      = GGML_TYPE_F16;
+                read_global_type = false;
+            }
+            tensor_storage.type = GGML_TYPE_F16;
+        }
+    }
+
+    void read_string(const std::string& str, zip_t* zip, std::string dir) {
+        if (str == "storage") {
+            read_global_type = true;
+        } else if (str != "state_dict") {
+            if (phase == READ_DATA) {
+                std::string entry_name = dir + "data/" + std::string(str);
+
+                size_t i, n = zip_entries_total(zip);
+                for (i = 0; i < n; ++i) {
+                    zip_entry_openbyindex(zip, i);
+                    {
+                        std::string name = zip_entry_name(zip);
+                        if (name == entry_name) {
+                            tensor_storage.index_in_zip = (int)i;
+                            entry_size                  = zip_entry_size(zip);
+                            zip_entry_close(zip);
+                            break;
+                        }
+                    }
+                    zip_entry_close(zip);
+                }
+
+                phase = entry_size > 0 ? CHECK_SIZE : READ_NAME;
+            }
+            if (!read_global_type && phase == READ_NAME) {
+                tensor_storage.name = str;
+                phase               = READ_DATA;
+                tensor_storage.type = global_type;
+            }
+        }
+    }
+};
+
+ggml_type PickleTensorReader::global_type = GGML_TYPE_F32;  // all pickle_tensors data type
+bool PickleTensorReader::read_global_type = false;
+
+static int find_char(uint8_t* buffer, int len, char c) {
+    for (int pos = 0; pos < len; pos++) {
+        if (buffer[pos] == c) {
+            return pos;
+        }
+    }
+    return -1;
+}
+
+static bool parse_data_pkl(uint8_t* buffer,
+                           size_t buffer_size,
+                           zip_t* zip,
+                           std::string dir,
+                           std::vector<TensorStorage>& tensor_storages,
+                           std::string* error) {
+    uint8_t* buffer_end = buffer + buffer_size;
+    if (buffer[0] == 0x80) {  // proto
+        if (buffer[1] != 2) {
+            set_error(error, "unsupported pickle protocol");
+            return false;
+        }
+        buffer += 2;  // 0x80 and version
+        char string_buffer[MAX_STRING_BUFFER];
+        bool finish = false;
+        PickleTensorReader reader;
+        // read pickle binary file
+        while (!finish && buffer < buffer_end) {
+            uint8_t opcode = *buffer;
+            buffer++;
+            // https://github.com/python/cpython/blob/3.7/Lib/pickletools.py#L1048
+            // https://github.com/python/cpython/blob/main/Lib/pickle.py#L105
+            switch (opcode) {
+                case '}':  // EMPTY_DICT     = b'}'   # push empty dict
+                    break;
+                case ']':  // EMPTY_LIST     = b']'   # push empty list
+                    break;
+                // skip unused sections
+                case 'h':  // BINGET         = b'h'   #   "    "    "    "   "   "  ;   "    " 1-byte arg
+                case 'q':  // BINPUT         = b'q'   #   "     "    "   "   " ;   "    " 1-byte arg
+                case 'Q':  // BINPERSID      = b'Q'   #  "       "         "  ;  "  "   "     "  stack
+                    buffer++;
+                    break;
+                case 'r':  // LONG_BINPUT    = b'r'   #   "     "    "   "   " ;   "    " 4-byte arg
+                    buffer += 4;
+                    break;
+                case 0x95:  // FRAME            = b'\x95'  # indicate the beginning of a new frame
+                    buffer += 8;
+                    break;
+                case 0x94:  // MEMOIZE          = b'\x94'  # store top of the stack in memo
+                    break;
+                case '(':  // MARK           = b'('   # push special markobject on stack
+                    break;
+                case 'K':  // BININT1        = b'K'   # push 1-byte unsigned int
+                {
+                    uint8_t value = *buffer;
+                    if (reader.read_int_value(value)) {
+                        buffer++;
+                    }
+                    buffer++;
+                } break;
+                case 'M':  // BININT2        = b'M'   # push 2-byte unsigned int
+                {
+                    uint16_t value = read_short(buffer);
+                    if (reader.read_int_value(value)) {
+                        buffer++;
+                    }
+                    buffer += 2;
+                } break;
+                case 'J':  // BININT         = b'J'   # push four-byte signed int
+                {
+                    const int32_t value = read_int(buffer);
+                    if (reader.read_int_value(value)) {
+                        buffer++;  // skip tuple after read num_elements
+                    }
+                    buffer += 4;
+                } break;
+                case 'X':  // BINUNICODE     = b'X'   #   "     "       "  ; counted UTF-8 string argument
+                {
+                    const int32_t len = read_int(buffer);
+                    buffer += 4;
+                    memset(string_buffer, 0, MAX_STRING_BUFFER);
+                    if (len > MAX_STRING_BUFFER) {
+                        // keep truncated names null-terminated, matching the old parser behavior
+                    }
+                    memcpy(string_buffer, buffer, len < MAX_STRING_BUFFER ? len : (MAX_STRING_BUFFER - 1));
+                    buffer += len;
+                    reader.read_string(string_buffer, zip, dir);
+                } break;
+                case 0x8C:  // SHORT_BINUNICODE = b'\x8c'  # push short string; UTF-8 length < 256 bytes
+                {
+                    const int8_t len = *buffer;
+                    buffer++;
+                    memset(string_buffer, 0, MAX_STRING_BUFFER);
+                    memcpy(string_buffer, buffer, len);
+                    buffer += len;
+                    // printf("String: '%s'\n", string_buffer);
+                } break;
+                case 'c':  // GLOBAL         = b'c'   # push self.find_class(modname, name); 2 string args
+                {
+                    int len = find_char(buffer, MAX_STRING_BUFFER, '\n');
+
+                    buffer += len + 1;
+                    len = find_char(buffer, MAX_STRING_BUFFER, '\n');
+
+                    memset(string_buffer, 0, MAX_STRING_BUFFER);
+                    memcpy(string_buffer, buffer, len);
+                    buffer += len + 1;
+                    reader.read_global(string_buffer);
+                } break;
+                case 0x86:  // TUPLE2         = b'\x86'  # build 2-tuple from two topmost stack items
+                case 0x85:  // TUPLE1         = b'\x85'  # build 1-tuple from stack top
+                case 't':   // TUPLE          = b't'   # build tuple from topmost stack items
+                    if (reader.phase == PickleTensorReader::READ_DIMENS) {
+                        reader.tensor_storage.reverse_ne();
+                        tensor_storages.push_back(reader.tensor_storage);
+
+                        // LOG_DEBUG("%s", reader.tensor_storage.name.c_str());
+                        // reset
+                        reader = PickleTensorReader();
+                    }
+                    break;
+                case '.':  // STOP           = b'.'   # every pickle ends with STOP
+                    finish = true;
+                    break;
+                default:
+                    break;
+            }
+        }
+    }
+    return true;
+}
+
+bool read_ckpt_file(const std::string& file_path,
+                    std::vector<TensorStorage>& tensor_storages,
+                    std::string* error) {
+    zip_t* zip = zip_open(file_path.c_str(), 0, 'r');
+    if (zip == nullptr) {
+        set_error(error, "failed to open '" + file_path + "'");
+        return false;
+    }
+
+    tensor_storages.clear();
+    bool success = true;
+    int n        = (int)zip_entries_total(zip);
+    for (int i = 0; i < n; ++i) {
+        zip_entry_openbyindex(zip, i);
+        {
+            std::string name = zip_entry_name(zip);
+            size_t pos       = name.find("data.pkl");
+            if (pos != std::string::npos) {
+                std::string dir = name.substr(0, pos);
+                printf("ZIP %d, name = %s, dir = %s \n", i, name.c_str(), dir.c_str());
+                void* pkl_data = nullptr;
+                size_t pkl_size;
+                zip_entry_read(zip, &pkl_data, &pkl_size);
+
+                // LOG_DEBUG("%lld", pkl_size);
+
+                if (!parse_data_pkl((uint8_t*)pkl_data, pkl_size, zip, dir, tensor_storages, error)) {
+                    success = false;
+                }
+
+                free(pkl_data);
+            }
+        }
+        zip_entry_close(zip);
+
+        if (!success) {
+            break;
+        }
+    }
+    zip_close(zip);
+    return success;
+}
