+++ src/model_io/ckpt_io.h
@@ -1,0 +1,14 @@
+#ifndef __SD_MODEL_IO_CKPT_IO_H__
+#define __SD_MODEL_IO_CKPT_IO_H__
+
+#include <string>
+#include <vector>
+
+#include "tensor_storage.h"
+
+bool is_ckpt_file(const std::string& file_path);
+bool read_ckpt_file(const std::string& file_path,
+                    std::vector<TensorStorage>& tensor_storages,
+                    std::string* error = nullptr);
+
+#endif  // __SD_MODEL_IO_CKPT_IO_H__
